requirements.txt
APScheduler==3.10.4
emoji==2.12.1
greenlet==3.0.3
pyaes==1.6.1
pyasn1==0.6.0
python-decouple==3.8
pytz==2024.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.32
Telethon==1.36.0
typing_extensions==4.12.2
tzlocal==5.2
